+++ lib/src/widget/asset_picker_viewer.dart
@@ -43,7 +43,11 @@ class AssetPickerViewer<Asset, Path> extends StatefulWidget {
     AssetSelectPredicate<AssetEntity>? selectPredicate,
     PermissionRequestOption permissionRequestOption =
         const PermissionRequestOption(),
+    bool shouldAutoplayPreview = false,
   }) async {
+    if (previewAssets.isEmpty) {
+      throw StateError('Previewing empty assets is not allowed.');
+    }
     await AssetPicker.permissionCheck(requestOption: permissionRequestOption);
     final Widget viewer = AssetPickerViewer<AssetEntity, AssetPathEntity>(
       builder: DefaultAssetPickerViewerBuilderDelegate(
@@ -65,6 +69,7 @@ class AssetPickerViewer<Asset, Path> extends StatefulWidget {
         maxAssets: maxAssets,
         shouldReversePreview: shouldReversePreview,
         selectPredicate: selectPredicate,
+        shouldAutoplayPreview: shouldAutoplayPreview,
       ),
     );
     final PageRouteBuilder<List<AssetEntity>> pageRoute =