+++ CHANGELOG.md
@@ -7,18 +7,82 @@ that can be found in the LICENSE file. -->
 > [!IMPORTANT]  
 > See the [Migration Guide](guides/migration_guide.md) for the details of breaking changes between versions.
 
-## 9.0.5
+## Unreleased
+
+*None.*
+
+## 9.3.1
+
+### Improvements
+
+- Do not mute the Live Photo during the preview.
+
+### Fixes
+
+- Fix the GIF indicator's layout.
+- Fix the directionality with the reverted grid item.
+
+## 9.3.0
+
+### New features
+
+- Add explicit Live Photos indicator for assets.
+
+### Improvements
+
+- Make Live Photos gesture consistent when scaling and panning.
+- Integrate `LocallyAvailableBuilder` with thumbnail options to improve the thumbnail loading speed.
+- Use `visibility_detector` and scroll observer to improve media playing experiences.
+
+### Fixes
+
+- Fix the bottom actions bar display conditions.
+
+## 9.2.2
+
+- Uses correct `isOriginal` for the `LocallyAvailableBuilder`.
+
+## 9.2.1
+
+- Improve changes when limited on iOS.
+- Use `LocallyAvailableBuilder` in the grid to provide better user awareness.
+
+## 9.2.0
+
+### New features
+
+- Introduce `AssetsChangeCallback` and `AssetsChangeRefreshPredicate`
+  to help users act according to asset changes.
+- Add `shouldAutoplayPreview` to the picker config.
+
+### Fixes
+
+- Raise detailed negative range error.
+- Fix viewer confirm button predication.
+- Enlarge GIF gradients.
+- Fix potential paths assets count unexpected merging behaviors.
+
+## 9.1.0
+
+### Improvements
+
+- Support limited permission displays on Android.
+- Improves the limited overlay padding on Android.
+- Adds permission request lock for the picker state.
+- Speeding up by splitting asset loading into separate steps.
+- Speeding up using `AdvancedCustomFilter` rather than `FilterOptionGroup` by default.
 
 ### Fixes
 
 - Fix reverted index when previewing assets on Android.
+- Requests with the correct options with the picker.
 
 ## 9.0.4
 
 ### Fixes
 
 - Fix the app bar of the viewer that is not animating.
-- Fix loading when no assets in the path.
+- Fix loading when no assets are in the path.
 - Reset the has more to load flag between path switching.
 
 ## 9.0.3
@@ -37,7 +101,7 @@ that can be found in the LICENSE file. -->
 
 ### Fixes
 
-- Fix current asset in the picker viewer.
+- Fix the current asset in the picker viewer.
 
 ### Improvements